添加接口: sweetcat-app-credit微服务添加接口：兑换商品。 兑换商品过程包括： 1.扣减用户积分， 2.扣减商品库存， 3.增加积分收支记录， 4.增加用户兑换记录 尚未完善，待添加用户优惠券微服务时，需要增加用户优惠券等

+++ sweetcat-app-credit/src/main/java/com/sweetcat/credit/application/command/AddBaseCommodityCommand.java
@@ -40,7 +40,7 @@ public class AddBaseCommodityCommand {
     /**
      * 兑换商品所需积分数
      */
-    private Integer creditNumber;
+    private Long creditNumber;
 
     /**
      * 商品类别：0优惠券；1实物商品
