+++ sweetcat-app-credit/src/main/java/com/sweetcat/credit/infrastructure/po/BaseCommodityPO.java
@@ -50,7 +50,7 @@ public class BaseCommodityPO implements Serializable {
     /**
      * 兑换商品所需积分数
      */
-    private Integer creditNumber;
+    private Long creditNumber;
 
     /**
      * 商品类别：0优惠券；1实物商品
