+++ sweetcat-app-credit/src/main/java/com/sweetcat/credit/infrastructure/po/CreditLogPO.java
@@ -1,7 +1,8 @@
 package com.sweetcat.credit.infrastructure.po;
 
-import java.io.Serializable;
 import lombok.Data;
+
+import java.io.Serializable;
 
 /**
  * t_app_credit_log
@@ -32,7 +33,7 @@ public class CreditLogPO implements Serializable {
     /**
      * 纪录涉及的积分数量
      */
-    private Integer creditNumber;
+    private Long creditNumber;
 
     /**
      * 记录创建时间
