+++ sweetcat-app-credit/src/main/java/com/sweetcat/credit/interfaces/facade/restdto/CreditLogRestDTO.java
@@ -36,7 +36,7 @@ public class CreditLogRestDTO implements Serializable {
     /**
      * 纪录涉及的积分数量
      */
-    private Integer creditNumber;
+    private Long creditNumber;
     /**
      * 记录创建时间
      */
